outsource clone from unittests
