+++ tests/active_workflow_sortindex_updown_test.php
@@ -18,9 +18,9 @@ defined('MOODLE_INTERNAL') || die();
 
 require_once(__DIR__ . '/generator/lib.php');
 require_once(__DIR__ . '/../lib.php');
+require_once(__DIR__ . '/workflow_actions_test_setup.php');
 
 use tool_lifecycle\manager\workflow_manager;
-use tool_lifecycle\entity\workflow;
 
 /**
  * Tests the different state changes of the workflow sortindex for up and down action.
@@ -31,68 +31,7 @@ use tool_lifecycle\entity\workflow;
  * @copyright  2017 Tobias Reischmann WWU
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class tool_lifecycle_workflow_sortindex_updown_testcase extends \advanced_testcase {
-
-    private $workflow1;
-    private $workflow2;
-    private $workflow3;
-
-    public function setUp() {
-        $this->resetAfterTest(true);
-        $generator = $this->getDataGenerator()->get_plugin_generator('tool_lifecycle');
-
-        // Remove preset workflows.
-        $workflows = workflow_manager::get_active_automatic_workflows();
-        foreach ($workflows as $workflow) {
-            workflow_manager::remove($workflow->id, true); // remove() hasn't removed unremovable workflows (like presets) anymore…
-        }
-
-        $this->workflow1 = $generator->create_workflow();
-        $this->workflow2 = $generator->create_workflow();
-        $this->workflow3 = $generator->create_workflow();
-
-        $this->assertFalse($this->workflow1->active);
-        $this->assertFalse($this->workflow2->active);
-        $this->assertFalse($this->workflow3->active);
-        $this->assertNull($this->workflow1->sortindex);
-        $this->assertNull($this->workflow2->sortindex);
-        $this->assertNull($this->workflow3->sortindex);
-    }
-
-    /**
-     * Test to activate the first workflow.
-     */
-    public function test_activate_first() {
-        workflow_manager::handle_action(ACTION_WORKFLOW_ACTIVATE, $this->workflow1->id);
-        $reloadworkflow = workflow_manager::get_workflow($this->workflow1->id);
-        $this->assertTrue(workflow_manager::is_active($this->workflow1->id));
-        $this->assertEquals(1, $reloadworkflow->sortindex);
-    }
-
-    /**
-     * Test to activate the first and the second workflow.
-     */
-    public function test_activate_second() {
-        workflow_manager::handle_action(ACTION_WORKFLOW_ACTIVATE, $this->workflow1->id);
-        workflow_manager::handle_action(ACTION_WORKFLOW_ACTIVATE, $this->workflow2->id);
-
-        $reloadworkflow = workflow_manager::get_workflow($this->workflow2->id);
-        $this->assertTrue(workflow_manager::is_active($this->workflow2->id));
-        $this->assertEquals(2, $reloadworkflow->sortindex);
-    }
-
-    /**
-     * Test to activate all three workflow.
-     */
-    public function test_activate_third() {
-        workflow_manager::handle_action(ACTION_WORKFLOW_ACTIVATE, $this->workflow1->id);
-        workflow_manager::handle_action(ACTION_WORKFLOW_ACTIVATE, $this->workflow2->id);
-        workflow_manager::handle_action(ACTION_WORKFLOW_ACTIVATE, $this->workflow3->id);
-
-        $reloadworkflow = workflow_manager::get_workflow($this->workflow3->id);
-        $this->assertTrue(workflow_manager::is_active($this->workflow3->id));
-        $this->assertEquals(3, $reloadworkflow->sortindex);
-    }
+class tool_lifecycle_workflow_sortindex_updown_testcase extends workflow_actions_test_setup {
 
     /**
      * Test to put down the first workflow.
